+++ Njsast/Coverage/ITextFileReader.cs
@@ -1,0 +1,9 @@
+using System;
+
+namespace Njsast.Coverage
+{
+    public interface ITextFileReader
+    {
+        ReadOnlySpan<byte> ReadUtf8(string fileName);
+    }
+}