+++ CHANGELOG.md
@@ -2,11 +2,15 @@
 
 ## [unreleased]
 
+### Changed
+
+- Default TypeScript version is 3.8.3
+
 ### Added
 
-New coverage reporter json-details.
+- New coverage reporter json-details.
 
-New js command with globals subcommand to print all global variables from provided js file.
+- New js command with globals subcommand to print all global variables from provided js file.
 
 ```sh
 bb js globals file.js